libtoyos: discover pci-serial devices
This is necessary to run on more machines in our test infrastructure. The
serial setup now first checks if a PCI serial device is found. If not, it
falls back to the regular BDA approach.

This can be tested in QEMU like this:

qemu-system-x86_64 -machine q35,accel=kvm -kernel ./src/tests/pit-timer.elf32 \
  -m 24M -no-reboot --cpu host -chardev stdio,id=hostserial \
  -device pci-serial,chardev=hostserial