pit-timer: fix draining of PIC interrupts
This didn't work as intended on some machines. So, now in the prologue, we
always drain all interrupts. After that, we only drain the ones that are expected
to be there.

Specifically on our lewisburg machine, interrupt 35 (pin 3) was raised from time
to time in some runs. From the Intel Platform Controller Hub spec in
"Table 5-13 Interrupt controller connections" we can find that the serial device
is the root of all evil here ;) So just always drain all interrupts initially,
and we are good to go.