pit-timer: make test independent for unexpected unrelated interrupts in IRR
On some machines, such as lewisburg, we encountered rare conditions where the
serial device fires interrupts. They have a higher priority due to a higher
vector number. To work around this, we modify the test in a way that is only
looks that the PIT timer is not in IRR where it shouldn't be. It is overkill to
require IRR==0 in these cases.