Revert "lapic-priority: enable test case again that works in latest VBox/KVM"
This reverts commit a5775bfa0412e9acded7c6250b5e1f32e5a52f2d.

We decided to do this as currently we're having the problem that our Virtual Box
repo runs the guest tests in vanilla/native and kvm mode. However, the test only
succeeds with kvm. As a workaround, we go back to the old behavior for now and
come up with a better solution soon.