libtoyos: add console_debugcon for easier debugging
This commit adds an additional console using the QEMU debugcon device to print
information in the very early stages. This is useful to debug the parsing of
the cmdline itself. It is only active in virtual hardware where VMMs usually
just ignore writes to unknown ports, if they don't support it.

More information about the QEMU debugcon device can be found here: