treewide: remove more debugport legacy
Remove more legacy of the debugport from SVP, that we don't use
anymore. Most changes in this commit are in cpuid/main.cpp.

We don't remove debugport_interface.h entirely, as it is tightly
integrated into lots inline-assembly statements. As this feature
doesn't hurt, and may be revived in the future, we keep it as is for now.