cmake: test parsing of serial port
This commit shows the immense effort necessary to unit test our legacy code
base. In a perfect world, we would have a `libtoyos-generic` that is as
compatible as can be with the standard library, so that it can be easily
tested on the host, and a `libtoyos-baremetal` with the stuff that can't be
build for the host system.

But we must live with the technical debt now. This is the best effort to
unit test some of our code for now.

Yes, we have a header dependency hell here, there is no clear structure.
But luckily, we have a rewrite in Rust on the horizon, so we can live with
this technical debt for now.