cmake: extract header files from libc-tiny to new libgtbase
This is a follow-up for the CMake-based unit tests. There, I accidentally forgot
to add `toyos/console_serial_util.cpp` to the unit tests. Because of that, we
thought things are simpler than they are. Because we need files such as
compiler.h in our libcxx, libtoyos, and also in libtoyos-host, we need to put
these header files at a location where it can be included everywhere safely
without breaking the build. They are moved to libgtbase (gt: guest tests)