lapic-timer: always set handler in drain function
The current situation is a little unfortunate, because drain_periodic_timer_irqs()
expects the calling code to have an IRQ guard specified. This is not always the
case. For more correct code, we always set a handler inside this function.
Luckily, this doesn't cause any side-effects that we rely on in our testing.