lapic-timer: add callback to fire interrupt in wait_for_interrupts
With this we fight a flakiness in several virtualization stacks, such as Virtual
Box. By firing an interrupt in the wait_for_interrupt() function itself, we
prevent situations in that a guest test is preempted after writing the init count
but before unmasking the interrupt. While this doesn't happen on real hardware,
in preempted situations we are loosing interrupts without this change.