nix: split public and cyberus-related Nix code
We want to open-source this repository. But we do not only want to
open source it as is, but also enable some of our Nix infrastructure
to the public.

So, we add nix/public and nix/cyberus. nix/cyberus has additional
tooling that only works when `pkgs.cyberus`, aka our cbspkgs overlay,
is active. We need this mostly for our CI pipeline.