nix: init flake, drop niv
Init a Nix flake, which is the modern variant to package a Nix
project and manage its dependencies. niv is dropped.

We try to keep as much as possible of the old Nix
attribute structure (public API).

As Flake packages can't be nested attributes, we've added
a new "guest-tests" attribute which contains all the tests
in all binary variants.